fix the crash in trisim upon runid assignment while debugging Windows 64bit

+++ src/engines_gpl/flow2d3d/packages/flow2d3d/include/flow2d3d.h
@@ -66,7 +66,7 @@ extern "C" {
         int *   contextID,
         int *   fsmFlags,
         const char *  runID,
-        int     runIDLen
+        size_t  runIDLen
         );
     }
